.history/requirements_20201210090115.txt
asgiref==3.3.1
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.11.8
cloudinary==1.23.0
confusable-homoglyphs==3.2.0
coverage==5.3
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-registration==3.1.1
django-star-ratings==0.9.1
django-tinymce==3.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==2.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
